requirements.txt
# requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database (PostgreSQL)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
psycopg2-binary

# Prometheus metrics
prometheus-client>=0.17.0

# python-multipart for file uploads with FastAPI forms
python-multipart>=0.0.6

# S3/MinIO Object Storage
boto3>=1.28.0          # AWS SDK for Python (used by aioboto3 implicitly for some things)
aioboto3>=11.2.0       # Asynchronous AWS SDK

# Celery for asynchronous tasks
celery[redis]>=5.3.0    # Includes redis-py client
flower>=2.0.0           # Optional: Celery monitoring tool

passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0 # For FastAPI's OAuth2 support if not using PyJWT directly everywhere
PyJWT>=2.0.0 # For our custom JWT encode/decode in security.py

gunicorn>=21.2.0 # For running FastAPI in production

# For structured JSON logging
python-json-logger>=2.0.0
